+++ dzVents/runtime/device-adapters/counter_device.lua
@@ -5,7 +5,9 @@ return {
 	name = 'Counter device adapter',
 
 	matches = function (device, adapterManager)
-		local res = (device.deviceSubType == 'RFXMeter counter' or device.deviceSubType == 'Counter Incremental')
+		local res = (       device.deviceSubType == 'RFXMeter counter' 
+                        or  device.deviceSubType == 'Counter Incremental'
+                        or  device.deviceSubType == 'Managed Counter')
 
 		if (not res) then
 			adapterManager.addDummyMethod(device, 'updateCounter')
@@ -25,8 +27,11 @@ return {
 		valueFormatted = device.counter or ''
 		info = adapterManager.parseFormatted(valueFormatted, domoticz['radixSeparator'])
 		device['counter'] = info['value']
-
-
+        
+        if device.deviceSubType == 'Managed Counter' and ( device.valueUnits == nil or device.valueUnits == "" ) then
+            device.valueUnits = string.match(device._data.data.counter, "%a+%d*") or ''  
+        end
+        
 		function device.updateCounter(value)
 			return device.update(0, value)
 		end