+++ dzVents/runtime/dzVents.lua
@@ -39,7 +39,7 @@ if (tonumber(globalvariables['dzVents_log_level']) == utils.LOG_DEBUG or TESTMOD
 
 	local events, length = helpers.getEventSummary()
 	if (length > 0) then
-		print('Debug: dzVents version: 2.4.9')
+		print('Debug: dzVents version: 2.4.12')
 
 		print('Debug: Event triggers:')
 		for i, event in pairs(events) do